Reference the hosts primary_interface
When you have multiple machines with different interfaces you can't
assume that the primary_interface is the same everywhere. The default
undercloud_floating_ip uses the current hosts primary_interface value to
lookup something on another host. This fails when these are different so
use the primary_interface specific to that host when looking these up.