Fixup flavors for nova
This is part of a inflight merge conflict review.

It's inappropriate to use MySQL to update these flavors, and
unnecessary. Also since b69296eb74edab12de161fce50e3a65ea18172f0 in
Nova, they're not created by migrations anymore, so we need to make sure
they're present.

This fixes up the existing ironic specific flavor adding and makes it a
generic nova concept, as well as adding to the example environment.